refactor(workflow): split 2552-line workflowRoutes.js into per-concern modules
registerWorkflowRoutes registered ~22 endpoints across five unrelated
concerns (definition CRUD, execution lifecycle, SSE streaming, human
checkpoints, admin listing/toggle, version publish/activate) in one
function. Split into definitionRoutes.js, executionRoutes.js,
streamRoutes.js, adminRoutes.js, and versionRoutes.js, each registering
its own slice; shared stateless helpers (filterByPermissions, isAdmin,
loadWorkflows, findWorkflowFile, validateWorkflow, checkWorkflowsFeature,
workflowClients) move into workflowRouteHelpers.js. workflowRoutes.js is
now a thin aggregator that wires the shared WorkflowEngine and calls each
register*Routes function.

Pure refactor, no behavior change — verified by booting the server,
logging in as the local admin, and exercising list/create/execute/
get-state/cancel/archive/export/SSE-stream/toggle/admin-list endpoints
end-to-end.

Fixes #1773
