+++ server/routes/workflow/workflowRoutes.js
@@ -5,292 +5,22 @@
  * Includes support for workflow CRUD operations, execution control, and
  * real-time progress streaming via Server-Sent Events (SSE).
  *
+ * Route registration is split by concern across sibling modules:
+ * definitionRoutes (CRUD), executionRoutes (lifecycle), streamRoutes (SSE +
+ * human checkpoints), adminRoutes (admin listing/toggle), and versionRoutes
+ * (publish/activate). This file just wires the shared WorkflowEngine and
+ * registers each group.
+ *
  * @module routes/workflow/workflowRoutes
  */
 
-import { promises as fs } from 'fs';
-import { join } from 'path';
-import { authRequired } from '../../middleware/authRequired.js';
-import { adminAuth } from '../../middleware/adminAuth.js';
-import { buildServerPath } from '../../utils/basePath.js';
 import { WorkflowEngine } from '../../services/workflow/WorkflowEngine.js';
-import { getExecutionRegistry } from '../../services/workflow/ExecutionRegistry.js';
-import { HumanNodeExecutor } from '../../services/workflow/executors/HumanNodeExecutor.js';
-import { actionTracker } from '../../actionTracker.js';
-import { createSseChannel, startInactiveClientSweep } from '../../utils/sseChannel.js';
-import { workflowConfigSchema } from '../../validators/workflowConfigSchema.js';
-import { atomicWriteJSON } from '../../utils/atomicWrite.js';
-import { getRootDir } from '../../pathUtils.js';
-import {
-  isValidWorkflowVersion,
-  validateIdForPath,
-  resolveAndValidatePath,
-  resolveAndValidateRealPath
-} from '../../utils/pathSecurity.js';
-import {
-  sendNotFound,
-  sendBadRequest,
-  sendFailedOperationError,
-  sendInsufficientPermissions,
-  sendErrorResponse
-} from '../../utils/responseHelpers.js';
-import { filterResourcesByPermissions } from '../../utils/authorization.js';
-import logger from '../../utils/logger.js';
-import configCache from '../../configCache.js';
-import { requireFeature } from '../../featureRegistry.js';
-import { removeMarketplaceInstallation } from '../../utils/installationCleanup.js';
-
-const checkWorkflowsFeature = requireFeature('workflows');
-
-/**
- * SSE clients map for workflow execution streaming
- * Maps executionId -> { response, lastActivity }
- * @type {Map<string, {response: object, lastActivity: Date}>}
- */
-const workflowClients = new Map();
-startInactiveClientSweep(workflowClients, { component: 'WorkflowRoutes' });
-
-/**
- * Filters workflows based on user permissions from groups.json.
- * Uses the standard group-based permission system (permissions.workflows)
- * consistent with how apps, models, and prompts are handled.
- *
- * @param {Object[]} workflows - Array of workflow definitions
- * @param {Object} user - User object with groups and permissions
- * @returns {Object[]} Filtered array of accessible workflows
- */
-function filterByPermissions(workflows, user) {
-  if (!Array.isArray(workflows)) {
-    return [];
-  }
-
-  // Admin users can see all workflows
-  if (isAdmin(user)) {
-    return workflows;
-  }
-
-  // Use the standard permission system via user.permissions.workflows
-  const workflowPermissions = user?.permissions?.workflows;
-  if (!workflowPermissions) {
-    return [];
-  }
-
-  return filterResourcesByPermissions(workflows, workflowPermissions);
-}
-
-/**
- * Checks if a user has admin privileges.
- *
- * @param {Object} user - User object from request
- * @returns {boolean} True if user has admin access
- */
-function isAdmin(user) {
-  if (!user) return false;
-  return user.groups?.includes('admin') || user.permissions?.adminAccess === true;
-}
-
-/**
- * Maximum entries kept in the input preview shown on the executions list.
- */
-const INPUT_PREVIEW_MAX_ENTRIES = 6;
-/**
- * Per-value truncation length (characters) for string fields in the input preview.
- */
-const INPUT_PREVIEW_MAX_VALUE_LENGTH = 120;
-
-/**
- * Builds a sanitized input preview from a workflow execution's initial data.
- * Strips internal underscore-prefixed keys, drops empty values, truncates long
- * strings, and caps entry count. Object/array values are summarized.
- *
- * @param {Object} initialData - Raw initialData passed to /execute
- * @returns {Object|null} Preview object (possibly with a `__more` count), or null if empty
- */
-function buildInputPreview(initialData) {
-  if (!initialData || typeof initialData !== 'object') return null;
-
-  const preview = {};
-  const allKeys = Object.keys(initialData).filter(k => !k.startsWith('_'));
-  let kept = 0;
-
-  for (const key of allKeys) {
-    if (kept >= INPUT_PREVIEW_MAX_ENTRIES) break;
-    const value = initialData[key];
-    if (value === null || value === undefined || value === '') continue;
-
-    if (typeof value === 'string') {
-      preview[key] =
-        value.length > INPUT_PREVIEW_MAX_VALUE_LENGTH
-          ? value.slice(0, INPUT_PREVIEW_MAX_VALUE_LENGTH - 1) + '…'
-          : value;
-    } else if (typeof value === 'number' || typeof value === 'boolean') {
-      preview[key] = value;
-    } else if (Array.isArray(value)) {
-      preview[key] = `[${value.length} items]`;
-    } else if (typeof value === 'object') {
-      const keys = Object.keys(value);
-      preview[key] =
-        keys.length === 0
-          ? '{}'
-          : `{${keys.slice(0, 4).join(', ')}${keys.length > 4 ? ', …' : ''}}`;
-    } else {
-      preview[key] = String(value);
-    }
-    kept += 1;
-  }
-
-  if (kept === 0) return null;
-
-  const remaining =
-    allKeys.filter(
-      k => initialData[k] !== null && initialData[k] !== undefined && initialData[k] !== ''
-    ).length - kept;
-  if (remaining > 0) preview.__more = remaining;
-
-  return preview;
-}
-
-/**
- * Builds the distinct list of model IDs referenced by a workflow definition's
- * nodes. Used at registration time so the executions list can show which
- * model(s) a run uses.
- *
- * @param {Object} workflow - Workflow definition
- * @returns {string[]} Distinct model IDs in insertion order
- */
-function buildModelsList(workflow) {
-  const set = new Set();
-  for (const node of workflow?.nodes || []) {
-    if (typeof node.model === 'string' && node.model) set.add(node.model);
-  }
-  return Array.from(set);
-}
-
-/**
- * Loads all workflow definitions from the filesystem.
- * Workflows are stored as individual JSON files in contents/workflows/
- *
- * @param {boolean} includeDisabled - Whether to include disabled workflows
- * @returns {Promise<Object[]>} Array of workflow definitions
- */
-async function loadWorkflows(includeDisabled = false) {
-  const rootDir = getRootDir();
-  const workflowsDir = join(rootDir, 'contents', 'workflows');
-
-  try {
-    // Create directory if it doesn't exist
-    await fs.mkdir(workflowsDir, { recursive: true });
-
-    const files = await fs.readdir(workflowsDir);
-    const jsonFiles = files.filter(f => f.endsWith('.json'));
-
-    const workflows = [];
-    for (const file of jsonFiles) {
-      try {
-        const filePath = join(workflowsDir, file);
-        const content = await fs.readFile(filePath, 'utf8');
-        const workflow = JSON.parse(content);
-
-        // Skip disabled workflows unless explicitly requested
-        if (!includeDisabled && workflow.enabled === false) {
-          continue;
-        }
-
-        workflows.push(workflow);
-      } catch (error) {
-        logger.warn('Failed to load workflow file', {
-          component: 'WorkflowRoutes',
-          file,
-          error: error.message
-        });
-      }
-    }
-
-    return workflows;
-  } catch (error) {
-    logger.error('Failed to read workflows directory', {
-      component: 'WorkflowRoutes',
-      error: error.message
-    });
-    return [];
-  }
-}
-
-/**
- * Finds the actual filename for a workflow ID.
- * Handles cases where the filename doesn't match the workflow ID.
- *
- * @param {string} workflowId - The workflow ID to search for
- * @param {string} workflowsDir - The workflows directory path
- * @returns {Promise<string|null>} The filename if found, null otherwise
- */
-async function findWorkflowFile(workflowId, workflowsDir) {
-  try {
-    const files = await fs.readdir(workflowsDir);
-    const jsonFiles = files.filter(f => f.endsWith('.json'));
-
-    // First try the expected filename
-    const expectedFilename = `${workflowId}.json`;
-    if (jsonFiles.includes(expectedFilename)) {
-      return expectedFilename;
-    }
-
-    // If not found, search through all files to find one with matching ID
-    for (const file of jsonFiles) {
-      try {
-        const filePath = join(workflowsDir, file);
-        const content = await fs.readFile(filePath, 'utf8');
-        const workflow = JSON.parse(content);
-        if (workflow.id === workflowId) {
-          return file;
-        }
-      } catch (error) {
-        // Skip files that can't be read or parsed
-        logger.debug('Skipping malformed workflow file', {
-          component: 'WorkflowRoutes',
-          file,
-          error: error.message
-        });
-      }
-    }
-
-    return null;
-  } catch (error) {
-    logger.warn('Failed to read workflows directory', {
-      component: 'WorkflowRoutes',
-      workflowsDir,
-      error: error.message
-    });
-    return null;
-  }
-}
-
-/**
- * Validates a workflow definition against the schema.
- *
- * @param {Object} workflow - Workflow definition to validate
- * @returns {{success: boolean, errors?: Object[]}} Validation result
- */
-function validateWorkflow(workflow) {
-  try {
-    const result = workflowConfigSchema.safeParse(workflow);
-    if (result.success) {
-      return { success: true, data: result.data };
-    }
-    return {
-      success: false,
-      errors: result.error.errors.map(err => ({
-        path: err.path.join('.'),
-        message: err.message
-      }))
-    };
-  } catch (error) {
-    return {
-      success: false,
-      errors: [{ path: '', message: error.message }]
-    };
-  }
-}
+import registerDefinitionRoutes from './definitionRoutes.js';
+import registerExecutionRoutes from './executionRoutes.js';
+import registerStreamRoutes from './streamRoutes.js';
+import registerAdminRoutes from './adminRoutes.js';
+import registerVersionRoutes from './versionRoutes.js';
+import { workflowClients } from './workflowRouteHelpers.js';
 
 /**
  * Registers all workflow-related API routes.
@@ -302,2173 +32,13 @@ function validateWorkflow(workflow) {
  */
 export default function registerWorkflowRoutes(app, deps = {}) {
   const workflowEngine = deps.workflowEngine || new WorkflowEngine();
-
-  // Recover persisted executions from disk on startup
-  const registry = getExecutionRegistry();
-  (async () => {
-    try {
-      await registry.loadFromDisk();
-      // Mark previously-running executions as failed (server process died)
-      for (const exec of registry.getActive()) {
-        if (exec.status === 'running') {
-          registry.updateStatus(exec.executionId, 'failed', { currentNode: null });
-        }
-      }
-    } catch (error) {
-      logger.error('Failed to load execution registry from disk', {
-        component: 'WorkflowRoutes',
-        error: error.message
-      });
-    }
-  })();
-
-  // ============================================================================
-  // Workflow Definition Endpoints
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows:
-   *   get:
-   *     summary: List available workflows
-   *     description: |
-   *       Retrieves all workflow definitions that the authenticated user has permission to access.
-   *       Workflows are filtered based on the user's group memberships and the workflow's allowedGroups.
-   *
-   *       **Permission Filtering:**
-   *       - Workflows without allowedGroups are accessible to all authenticated users
-   *       - Workflows with allowedGroups are only shown to users in those groups
-   *       - Admin users can see all workflows
-   *     tags:
-   *       - Workflows
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     responses:
-   *       200:
-   *         description: List of accessible workflows
-   *         content:
-   *           application/json:
-   *             schema:
-   *               type: array
-   *               items:
-   *                 type: object
-   *       401:
-   *         description: Authentication required
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const workflows = await loadWorkflows();
-        const filtered = filterByPermissions(workflows, req.user);
-
-        res.json(filtered);
-      } catch (error) {
-        sendFailedOperationError(res, 'fetch workflows', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/my-executions:
-   *   get:
-   *     summary: List user's workflow executions
-   *     description: |
-   *       Retrieves all workflow executions started by the current user.
-   *       Supports filtering by status and pagination.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: query
-   *         name: status
-   *         schema:
-   *           type: string
-   *           enum: [pending, running, paused, completed, failed, cancelled]
-   *         description: Filter by execution status
-   *       - in: query
-   *         name: limit
-   *         schema:
-   *           type: integer
-   *           default: 20
-   *         description: Maximum number of results
-   *       - in: query
-   *         name: offset
-   *         schema:
-   *           type: integer
-   *           default: 0
-   *         description: Number of results to skip
-   *     responses:
-   *       200:
-   *         description: List of user's executions
-   *       401:
-   *         description: Authentication required
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows/my-executions'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const userId = req.user?.id || req.user?.sub || req.user?.username || 'anonymous';
-        const { status, limit = 20, offset = 0, includeArchived } = req.query;
-
-        const registry = getExecutionRegistry();
-        const executions = registry.getByUser(userId, {
-          status,
-          includeArchived: includeArchived === 'true' || includeArchived === '1',
-          limit: parseInt(limit, 10),
-          offset: parseInt(offset, 10)
-        });
-
-        res.json(executions);
-      } catch (error) {
-        sendFailedOperationError(res, 'fetch user executions', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/{id}:
-   *   get:
-   *     summary: Get workflow definition
-   *     description: |
-   *       Retrieves a single workflow definition by ID.
-   *       User must have permission to access the workflow.
-   *     tags:
-   *       - Workflows
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         description: Unique identifier of the workflow
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Workflow definition
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Access denied
-   *       404:
-   *         description: Workflow not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows/:id'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { id } = req.params;
-
-        // Validate ID for security
-        if (!validateIdForPath(id, 'workflow', res)) {
-          return;
-        }
-
-        const workflows = await loadWorkflows();
-        const workflow = workflows.find(w => w.id === id);
-
-        if (!workflow) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        // Check if user has permission to access this workflow
-        const accessible = filterByPermissions([workflow], req.user);
-        if (accessible.length === 0) {
-          return sendInsufficientPermissions(res, 'workflow access');
-        }
-
-        res.json(workflow);
-      } catch (error) {
-        sendFailedOperationError(res, 'fetch workflow', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows:
-   *   post:
-   *     summary: Create workflow (admin only)
-   *     description: |
-   *       Creates a new workflow definition.
-   *       Requires admin privileges.
-   *
-   *       **Validation:**
-   *       - Workflow must pass schema validation
-   *       - Workflow ID must be unique
-   *     tags:
-   *       - Workflows
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     requestBody:
-   *       required: true
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *     responses:
-   *       201:
-   *         description: Workflow created successfully
-   *       400:
-   *         description: Validation error
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Admin access required
-   *       409:
-   *         description: Workflow with this ID already exists
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/workflows'),
-    checkWorkflowsFeature,
-    adminAuth,
-    async (req, res) => {
-      try {
-        const workflowData = req.body;
-
-        // Validate workflow schema
-        const validation = validateWorkflow(workflowData);
-        if (!validation.success) {
-          return sendBadRequest(res, 'Invalid workflow definition', validation.errors);
-        }
-
-        // Validate ID for security
-        if (!validateIdForPath(workflowData.id, 'workflow', res)) {
-          return;
-        }
-
-        // Check if workflow already exists
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-        await fs.mkdir(workflowsDir, { recursive: true });
-
-        const existingFile = await findWorkflowFile(workflowData.id, workflowsDir);
-        if (existingFile) {
-          return sendErrorResponse(res, 409, 'Workflow with this ID already exists');
-        }
-
-        // Write workflow file
-        const workflowPath = join(workflowsDir, `${workflowData.id}.json`);
-        await atomicWriteJSON(workflowPath, validation.data);
-
-        logger.info('Workflow created', {
-          component: 'WorkflowRoutes',
-          workflowId: workflowData.id
-        });
-
-        res.status(201).json({
-          message: 'Workflow created successfully',
-          workflow: validation.data
-        });
-
-        // Refresh workflow cache after successful creation
-        configCache.refreshWorkflowsCache?.();
-      } catch (error) {
-        sendFailedOperationError(res, 'create workflow', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/{id}:
-   *   put:
-   *     summary: Update workflow (admin only)
-   *     description: |
-   *       Updates an existing workflow definition.
-   *       Requires admin privileges.
-   *
-   *       **Validation:**
-   *       - Workflow must pass schema validation
-   *       - Workflow ID cannot be changed
-   *     tags:
-   *       - Workflows
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *     requestBody:
-   *       required: true
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *     responses:
-   *       200:
-   *         description: Workflow updated successfully
-   *       400:
-   *         description: Validation error
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Admin access required
-   *       404:
-   *         description: Workflow not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.put(
-    buildServerPath('/api/workflows/:id'),
-    checkWorkflowsFeature,
-    adminAuth,
-    async (req, res) => {
-      try {
-        const { id } = req.params;
-        const workflowData = req.body;
-
-        // Validate ID for security
-        if (!validateIdForPath(id, 'workflow', res)) {
-          return;
-        }
-
-        // Validate workflow schema
-        const validation = validateWorkflow(workflowData);
-        if (!validation.success) {
-          return sendBadRequest(res, 'Invalid workflow definition', validation.errors);
-        }
-
-        // Ensure ID cannot be changed
-        if (workflowData.id !== id) {
-          return sendBadRequest(res, 'Workflow ID cannot be changed');
-        }
-
-        // Find existing workflow file
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-        const filename = await findWorkflowFile(id, workflowsDir);
-
-        if (!filename) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        // Update workflow file
-        const workflowPath = join(workflowsDir, filename);
-        await atomicWriteJSON(workflowPath, validation.data);
-
-        logger.info('Workflow updated', {
-          component: 'WorkflowRoutes',
-          workflowId: id
-        });
-
-        res.json({
-          message: 'Workflow updated successfully',
-          workflow: validation.data
-        });
-
-        // Refresh workflow cache after successful update
-        configCache.refreshWorkflowsCache?.();
-      } catch (error) {
-        sendFailedOperationError(res, 'update workflow', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/{id}:
-   *   delete:
-   *     summary: Delete workflow (admin only)
-   *     description: |
-   *       Permanently deletes a workflow definition.
-   *       Requires admin privileges.
-   *
-   *       **Warning:** This operation cannot be undone.
-   *     tags:
-   *       - Workflows
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Workflow deleted successfully
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Admin access required
-   *       404:
-   *         description: Workflow not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.delete(
-    buildServerPath('/api/workflows/:id'),
-    checkWorkflowsFeature,
-    adminAuth,
-    async (req, res) => {
-      try {
-        const { id } = req.params;
-
-        // Validate ID for security
-        if (!validateIdForPath(id, 'workflow', res)) {
-          return;
-        }
-
-        // Find workflow file
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-        const filename = await findWorkflowFile(id, workflowsDir);
-
-        if (!filename) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        // Delete workflow file
-        const workflowPath = join(workflowsDir, filename);
-        await fs.unlink(workflowPath);
-
-        logger.info('Workflow deleted', {
-          component: 'WorkflowRoutes',
-          workflowId: id
-        });
-
-        res.json({
-          message: 'Workflow deleted successfully'
-        });
-
-        // Refresh workflow cache after successful deletion
-        configCache.refreshWorkflowsCache?.();
-        await removeMarketplaceInstallation('workflow', id);
-      } catch (error) {
-        sendFailedOperationError(res, 'delete workflow', error);
-      }
-    }
-  );
-
-  // ============================================================================
-  // Workflow Execution Endpoints
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows/{id}/execute:
-   *   post:
-   *     summary: Start workflow execution
-   *     description: |
-   *       Starts a new execution of the specified workflow.
-   *       Returns an execution ID that can be used to monitor progress.
-   *
-   *       **Initial Data:**
-   *       The initialData object is passed to the workflow as starting context.
-   *       It can contain any data needed by the workflow's nodes.
-   *
-   *       **Options:**
-   *       - checkpointOnNode: Create checkpoint after each node execution
-   *       - timeout: Override default node execution timeout
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *     requestBody:
-   *       required: false
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *             properties:
-   *               initialData:
-   *                 type: object
-   *                 description: Initial data/context for the workflow
-   *               options:
-   *                 type: object
-   *                 properties:
-   *                   checkpointOnNode:
-   *                     type: boolean
-   *                   timeout:
-   *                     type: number
-   *     responses:
-   *       200:
-   *         description: Execution started
-   *         content:
-   *           application/json:
-   *             schema:
-   *               type: object
-   *               properties:
-   *                 executionId:
-   *                   type: string
-   *                 status:
-   *                   type: string
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Access denied
-   *       404:
-   *         description: Workflow not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/workflows/:id/execute'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { id } = req.params;
-        const { initialData = {}, options = {} } = req.body || {};
-
-        // Validate ID for security
-        if (!validateIdForPath(id, 'workflow', res)) {
-          return;
-        }
-
-        // Load and validate workflow access
-        const workflows = await loadWorkflows();
-        const workflow = workflows.find(w => w.id === id);
-
-        if (!workflow) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        // Check if user has permission to execute this workflow
-        const accessible = filterByPermissions([workflow], req.user);
-        if (accessible.length === 0) {
-          return sendInsufficientPermissions(res, 'workflow execution');
-        }
-
-        // Store workflow definition in initial data for resume support
-        const enrichedInitialData = {
-          ...initialData,
-          _workflowDefinition: workflow
-        };
-
-        // Start workflow execution
-        const state = await workflowEngine.start(workflow, enrichedInitialData, {
-          ...options,
-          user: req.user
-        });
-
-        // Register with ExecutionRegistry for tracking
-        const userId = req.user?.id || req.user?.username || 'anonymous';
-        const registry = getExecutionRegistry();
-        registry.register(state.executionId, {
-          userId,
-          workflowId: workflow.id,
-          workflowName: workflow.name,
-          status: state.status,
-          startedAt: state.createdAt,
-          inputPreview: buildInputPreview(initialData),
-          models: buildModelsList(workflow)
-        });
-
-        logger.info('Workflow execution started', {
-          component: 'WorkflowRoutes',
-          workflowId: id,
-          executionId: state.executionId,
-          userId: req.user?.id
-        });
-
-        res.json({
-          executionId: state.executionId,
-          status: state.status,
-          workflowId: workflow.id,
-          startedAt: state.createdAt
-        });
-      } catch (error) {
-        // Handle specific workflow errors
-        if (error.code === 'WORKFLOW_CYCLE_DETECTED' || error.code === 'WORKFLOW_NO_START_NODE') {
-          return sendBadRequest(res, error.message);
-        }
-
-        sendFailedOperationError(res, 'start workflow execution', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}:
-   *   get:
-   *     summary: Get execution state
-   *     description: |
-   *       Retrieves the current state of a workflow execution.
-   *       Returns detailed information including status, current nodes,
-   *       completed nodes, execution history, and any errors.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Execution state
-   *       401:
-   *         description: Authentication required
-   *       404:
-   *         description: Execution not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows/executions/:executionId'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        const state = await workflowEngine.getState(executionId);
-
-        if (!state) {
-          return sendNotFound(res, 'Execution');
-        }
-
-        // Determine if execution can be reconnected (streaming)
-        const canReconnect = state.status === 'running' || state.status === 'paused';
-
-        // Get pending human checkpoint if any
-        const pendingCheckpoint = state.data?.pendingCheckpoint || null;
-
-        // Get workflow name if available
-        const workflowName = state.data?._workflowDefinition?.name || null;
-
-        // Return state with reconnection info
-        res.json({
-          executionId: state.executionId,
-          workflowId: state.workflowId,
-          workflowName,
-          status: state.status,
-          currentNodes: state.currentNodes,
-          completedNodes: state.completedNodes,
-          failedNodes: state.failedNodes,
-          createdAt: state.createdAt,
-          startedAt: state.startedAt,
-          completedAt: state.completedAt,
-          history: state.history,
-          errors: state.errors,
-          checkpoints: state.checkpoints?.map(cp => ({
-            id: cp.id,
-            timestamp: cp.timestamp
-          })),
-          // Include full data for node results display
-          data: state.data,
-          // Reconnection info
-          canReconnect,
-          pendingCheckpoint
-        });
-      } catch (error) {
-        sendFailedOperationError(res, 'fetch execution state', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}/resume:
-   *   post:
-   *     summary: Resume paused workflow
-   *     description: |
-   *       Resumes a paused workflow execution.
-   *       Can optionally include data to merge into the workflow context,
-   *       such as human responses for human-in-the-loop nodes.
-   *
-   *       **Resume Data:**
-   *       - humanResponse: Response from human-in-the-loop interaction
-   *       - Additional context data as needed
-   *
-   *       **Checkpoint:**
-   *       If checkpointId is provided, execution resumes from that checkpoint
-   *       instead of the current state.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     requestBody:
-   *       required: false
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *             properties:
-   *               humanResponse:
-   *                 description: Response data for human-in-the-loop nodes
-   *               checkpointId:
-   *                 type: string
-   *                 description: Optional checkpoint ID to resume from
-   *     responses:
-   *       200:
-   *         description: Execution resumed
-   *       400:
-   *         description: Cannot resume - invalid state
-   *       401:
-   *         description: Authentication required
-   *       404:
-   *         description: Execution not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/workflows/executions/:executionId/resume'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-        const { humanResponse, checkpointId, ...additionalData } = req.body || {};
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        // Build resume data
-        const resumeData = {
-          ...additionalData
-        };
-
-        if (humanResponse !== undefined) {
-          resumeData._humanResponse = humanResponse;
-        }
-
-        // Resume options
-        const options = {
-          user: req.user
-        };
-
-        if (checkpointId) {
-          options.checkpointId = checkpointId;
-        }
-
-        const state = await workflowEngine.resume(executionId, resumeData, options);
-
-        logger.info('Workflow execution resumed', {
-          component: 'WorkflowRoutes',
-          executionId,
-          userId: req.user?.id
-        });
-
-        res.json({
-          executionId: state.executionId,
-          status: state.status,
-          currentNodes: state.currentNodes
-        });
-      } catch (error) {
-        // Handle specific workflow errors
-        if (
-          error.code === 'EXECUTION_NOT_FOUND' ||
-          error.code === 'INVALID_STATE_FOR_RESUME' ||
-          error.code === 'WORKFLOW_NOT_AVAILABLE'
-        ) {
-          if (error.code === 'EXECUTION_NOT_FOUND') {
-            return sendNotFound(res, 'Execution');
-          }
-          return sendBadRequest(res, error.message);
-        }
-
-        sendFailedOperationError(res, 'resume workflow execution', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}/restart:
-   *   post:
-   *     summary: Restart a cancelled or failed execution
-   *     description: |
-   *       Resumes a workflow that was cancelled by timeout/server restart or
-   *       failed mid-execution. Previously-completed nodes are not re-run;
-   *       only the interrupted nodes (those in `currentNodes`) are picked up.
-   *       User-cancelled executions are refused.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   */
-  app.post(
-    buildServerPath('/api/workflows/executions/:executionId/restart'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        const state = await workflowEngine.resumeFromTerminated(executionId, {
-          user: req.user
-        });
-
-        logger.info('Workflow execution restarted from terminated state', {
-          component: 'WorkflowRoutes',
-          executionId,
-          userId: req.user?.id
-        });
-
-        res.json({
-          executionId: state.executionId,
-          status: state.status,
-          currentNodes: state.currentNodes
-        });
-      } catch (error) {
-        if (error.code === 'EXECUTION_NOT_FOUND') {
-          return sendNotFound(res, 'Execution');
-        }
-        if (
-          error.code === 'INVALID_STATE_FOR_RESUME' ||
-          error.code === 'WORKFLOW_NOT_AVAILABLE' ||
-          error.code === 'NO_RESUME_POINT' ||
-          error.code === 'USER_CANCELLED'
-        ) {
-          return sendBadRequest(res, error.message);
-        }
-        sendFailedOperationError(res, 'restart workflow execution', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}/cancel:
-   *   post:
-   *     summary: Cancel execution
-   *     description: |
-   *       Cancels a running or paused workflow execution.
-   *       The execution status will be set to 'cancelled'.
-   *
-   *       **Note:** Already completed or cancelled executions cannot be cancelled.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     requestBody:
-   *       required: false
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *             properties:
-   *               reason:
-   *                 type: string
-   *                 description: Reason for cancellation
-   *     responses:
-   *       200:
-   *         description: Execution cancelled
-   *       401:
-   *         description: Authentication required
-   *       404:
-   *         description: Execution not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/workflows/executions/:executionId/cancel'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-        const { reason = 'user_cancelled' } = req.body || {};
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        const state = await workflowEngine.cancel(executionId, reason);
-
-        logger.info('Workflow execution cancelled', {
-          component: 'WorkflowRoutes',
-          executionId,
-          reason,
-          userId: req.user?.id
-        });
-
-        res.json({
-          success: true,
-          executionId: state.executionId,
-          status: state.status
-        });
-      } catch (error) {
-        if (error.code === 'EXECUTION_NOT_FOUND') {
-          return sendNotFound(res, 'Execution');
-        }
-
-        sendFailedOperationError(res, 'cancel workflow execution', error);
-      }
-    }
-  );
-
-  // ============================================================================
-  // Delete & Archive Endpoints
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}:
-   *   delete:
-   *     summary: Hard-delete an execution
-   *     description: |
-   *       Removes a workflow execution from the registry and deletes its
-   *       checkpoint files on disk. Owner-only (or admin). Refuses to delete
-   *       a running execution — cancel it first.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Execution deleted
-   *       403:
-   *         description: Not the owner and not an admin
-   *       404:
-   *         description: Execution not found
-   *       409:
-   *         description: Execution is still running
-   */
-  app.delete(
-    buildServerPath('/api/workflows/executions/:executionId'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        const registry = getExecutionRegistry();
-        const execution = registry.get(executionId);
-
-        if (!execution) {
-          return sendNotFound(res, 'Execution');
-        }
-
-        const currentUserId = req.user?.id || req.user?.sub || req.user?.username || 'anonymous';
-        if (execution.userId !== currentUserId && !isAdmin(req.user)) {
-          return sendInsufficientPermissions(res, 'delete execution');
-        }
-
-        if (execution.status === 'running') {
-          return res.status(409).json({
-            error: 'cannot_delete_running',
-            message: 'Cancel the workflow before deleting it.'
-          });
-        }
-
-        try {
-          await workflowEngine.deleteExecution(executionId);
-        } catch (error) {
-          if (error.code === 'EXECUTION_ACTIVE') {
-            return res.status(409).json({
-              error: 'cannot_delete_running',
-              message: 'Cancel the workflow before deleting it.'
-            });
-          }
-          throw error;
-        }
-
-        registry.remove(executionId);
-
-        logger.info('Workflow execution deleted', {
-          component: 'WorkflowRoutes',
-          executionId,
-          userId: currentUserId
-        });
-
-        res.json({ success: true });
-      } catch (error) {
-        sendFailedOperationError(res, 'delete execution', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}:
-   *   patch:
-   *     summary: Update execution metadata (currently archive/unarchive)
-   *     description: |
-   *       Toggle the archived flag on an execution. Owner-only (or admin).
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     requestBody:
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *             properties:
-   *               archived:
-   *                 type: boolean
-   *     responses:
-   *       200:
-   *         description: Updated execution metadata
-   *       403:
-   *         description: Not the owner and not an admin
-   *       404:
-   *         description: Execution not found
-   */
-  app.patch(
-    buildServerPath('/api/workflows/executions/:executionId'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-        const { archived } = req.body || {};
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        if (typeof archived !== 'boolean') {
-          return sendBadRequest(res, 'archived must be a boolean');
-        }
-
-        const registry = getExecutionRegistry();
-        const execution = registry.get(executionId);
-
-        if (!execution) {
-          return sendNotFound(res, 'Execution');
-        }
-
-        const currentUserId = req.user?.id || req.user?.sub || req.user?.username || 'anonymous';
-        if (execution.userId !== currentUserId && !isAdmin(req.user)) {
-          return sendInsufficientPermissions(res, 'update execution');
-        }
-
-        const updated = registry.setArchived(executionId, archived);
-
-        logger.info('Workflow execution archive toggled', {
-          component: 'WorkflowRoutes',
-          executionId,
-          archived,
-          userId: currentUserId
-        });
-
-        res.json(updated);
-      } catch (error) {
-        sendFailedOperationError(res, 'update execution', error);
-      }
-    }
-  );
-
-  // ============================================================================
-  // Export Endpoint
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}/export:
-   *   get:
-   *     summary: Export execution state as JSON
-   *     description: |
-   *       Downloads the complete workflow execution state as a JSON file.
-   *       Useful for debugging and analysis of workflow execution.
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Execution state JSON file
-   *         content:
-   *           application/json:
-   *             schema:
-   *               type: object
-   *       401:
-   *         description: Authentication required
-   *       404:
-   *         description: Execution not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows/executions/:executionId/export'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        const state = await workflowEngine.getState(executionId);
-
-        if (!state) {
-          return sendNotFound(res, 'Execution');
-        }
-
-        // Build export object with all relevant state data
-        const exportData = {
-          executionId: state.executionId,
-          workflowId: state.workflowId,
-          status: state.status,
-          data: state.data,
-          history: state.history,
-          completedNodes: state.completedNodes,
-          currentNodes: state.currentNodes,
-          failedNodes: state.failedNodes,
-          errors: state.errors,
-          checkpoints: state.checkpoints,
-          timestamps: {
-            created: state.createdAt,
-            started: state.startedAt,
-            completed: state.completedAt
-          },
-          exportedAt: new Date().toISOString()
-        };
-
-        // Build a useful filename. Execution IDs are prefixed `wf-exec-`, so
-        // strip the prefix before slicing — otherwise every run's filename
-        // ends in the same literal `wf-exec-`.
-        const shortId = executionId.replace(/^wf-exec-/, '').slice(0, 8) || executionId;
-        const workflowSlug = (state.workflowId || 'workflow').replace(/[^a-zA-Z0-9._-]/g, '_');
-        res.setHeader('Content-Type', 'application/json');
-        res.setHeader(
-          'Content-Disposition',
-          `attachment; filename="${workflowSlug}-${shortId}.json"`
-        );
-
-        res.json(exportData);
-
-        logger.info('Workflow execution exported', {
-          component: 'WorkflowRoutes',
-          executionId,
-          userId: req.user?.id
-        });
-      } catch (error) {
-        sendFailedOperationError(res, 'export execution state', error);
-      }
-    }
-  );
-
-  // ============================================================================
-  // SSE Streaming Endpoint
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}/stream:
-   *   get:
-   *     summary: SSE event stream for workflow progress
-   *     description: |
-   *       Establishes a Server-Sent Events connection to receive real-time
-   *       updates about workflow execution progress.
-   *
-   *       **Events:**
-   *       - workflow.start: Workflow execution started
-   *       - workflow.node.start: Node execution started
-   *       - workflow.node.complete: Node execution completed
-   *       - workflow.node.error: Node execution failed
-   *       - workflow.paused: Workflow paused (e.g., for human input)
-   *       - workflow.complete: Workflow completed successfully
-   *       - workflow.failed: Workflow execution failed
-   *       - workflow.cancelled: Workflow was cancelled
-   *       - workflow.checkpoint.saved: Checkpoint was saved
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *       - Streaming
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: SSE stream established
-   *         content:
-   *           text/event-stream:
-   *             schema:
-   *               type: string
-   *       401:
-   *         description: Authentication required
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows/executions/:executionId/stream'),
-    checkWorkflowsFeature,
-    authRequired,
-    (req, res) => {
-      const { executionId } = req.params;
-
-      const channel = createSseChannel({
-        req,
-        res,
-        id: executionId,
-        map: workflowClients,
-        component: 'WorkflowRoutes',
-        onClose: () => actionTracker.off('fire-sse', handleWorkflowEvent)
-      });
-
-      channel.send('connected', { executionId });
-
-      logger.info('SSE connection established for workflow execution', {
-        component: 'WorkflowRoutes',
-        executionId
-      });
-
-      // Define event types to listen for. The handler matches by prefix,
-      // so `workflow.node` covers `workflow.node.start/complete/error/retry`,
-      // and `workflow.subworkflow` covers start/complete events emitted by
-      // the planner executor.
-      const workflowEventTypes = [
-        'workflow.start',
-        'workflow.iteration',
-        'workflow.node',
-        'workflow.paused',
-        'workflow.human',
-        'workflow.complete',
-        'workflow.failed',
-        'workflow.cancelled',
-        'workflow.checkpoint',
-        'workflow.plan',
-        'workflow.subworkflow'
-      ];
-
-      /**
-       * Handler for workflow events from actionTracker
-       * @param {Object} eventData - Event data
-       */
-      const handleWorkflowEvent = eventData => {
-        // Only forward events for this execution
-        if (eventData.chatId !== executionId && eventData.executionId !== executionId) {
-          return;
-        }
-
-        // Extract event type from the event data
-        const eventType = eventData.event;
-
-        // Check if this is a workflow event we care about. Compare with a
-        // trailing dot so that allowing `workflow.node` does not also accept
-        // a hypothetical `workflow.nodes` event.
-        if (
-          !workflowEventTypes.some(type => eventType === type || eventType.startsWith(`${type}.`))
-        ) {
-          return;
-        }
-
-        // Log event details for debugging
-        logger.debug('Sending SSE event to client', {
-          component: 'WorkflowRoutes',
-          executionId,
-          eventType,
-          nodeId: eventData.nodeId,
-          iteration: eventData.iteration
-        });
-
-        const sent = channel.send(eventType, eventData);
-
-        // Log successful send for important events
-        if (sent && (eventType === 'workflow.node.complete' || eventType === 'workflow.complete')) {
-          logger.info('SSE event sent', {
-            component: 'WorkflowRoutes',
-            executionId,
-            eventType,
-            nodeId: eventData.nodeId
-          });
-        }
-      };
-
-      // Register event handler
-      actionTracker.on('fire-sse', handleWorkflowEvent);
-    }
-  );
-
-  // ============================================================================
-  // Human Checkpoint Response Endpoint
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows/executions/{executionId}/respond:
-   *   post:
-   *     summary: Respond to human checkpoint
-   *     description: |
-   *       Provides a response to a human checkpoint (approval/input) node.
-   *       This will resume the paused workflow with the provided response.
-   *
-   *       **Response Object:**
-   *       - checkpointId: ID of the checkpoint being responded to (required)
-   *       - response: Selected option value (required)
-   *       - data: Optional additional form data if the checkpoint has an inputSchema
-   *     tags:
-   *       - Workflows
-   *       - Execution
-   *       - Human Checkpoint
-   *     security:
-   *       - bearerAuth: []
-   *       - cookieAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: executionId
-   *         required: true
-   *         schema:
-   *           type: string
-   *     requestBody:
-   *       required: true
-   *       content:
-   *         application/json:
-   *           schema:
-   *             type: object
-   *             required:
-   *               - checkpointId
-   *               - response
-   *             properties:
-   *               checkpointId:
-   *                 type: string
-   *                 description: ID of the checkpoint being responded to
-   *               response:
-   *                 type: string
-   *                 description: Selected option value
-   *               data:
-   *                 type: object
-   *                 description: Additional form data
-   *     responses:
-   *       200:
-   *         description: Response accepted, workflow resumed
-   *         content:
-   *           application/json:
-   *             schema:
-   *               type: object
-   *               properties:
-   *                 success:
-   *                   type: boolean
-   *                 newStatus:
-   *                   type: string
-   *       400:
-   *         description: Invalid response or checkpoint not found
-   *       401:
-   *         description: Authentication required
-   *       404:
-   *         description: Execution not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/workflows/executions/:executionId/respond'),
-    checkWorkflowsFeature,
-    authRequired,
-    async (req, res) => {
-      try {
-        const { executionId } = req.params;
-        const { checkpointId, response, data } = req.body || {};
-
-        if (!validateIdForPath(executionId, 'executionId')) {
-          return sendBadRequest(res, 'Invalid executionId');
-        }
-
-        // Validate required fields
-        if (!checkpointId) {
-          return sendBadRequest(res, 'checkpointId is required');
-        }
-
-        if (!response) {
-          return sendBadRequest(res, 'response is required');
-        }
-
-        // Get current execution state
-        const state = await workflowEngine.getState(executionId);
-
-        if (!state) {
-          return sendNotFound(res, 'Execution');
-        }
-
-        // Verify execution is paused with a pending checkpoint
-        if (state.status !== 'paused') {
-          return sendBadRequest(res, `Cannot respond to execution with status: ${state.status}`);
-        }
-
-        // Check if there's a pending checkpoint
-        const pendingCheckpoint = state.data?.pendingCheckpoint;
-        if (!pendingCheckpoint) {
-          return sendBadRequest(res, 'No pending checkpoint for this execution');
-        }
-
-        // Verify checkpoint ID matches
-        if (pendingCheckpoint.id !== checkpointId) {
-          return sendBadRequest(res, 'Checkpoint ID does not match pending checkpoint');
-        }
-
-        // Get the workflow and node for resuming
-        const workflow = state.data?._workflowDefinition;
-        if (!workflow) {
-          return sendBadRequest(res, 'Workflow definition not available for resume');
-        }
-
-        const humanNode = workflow.nodes.find(n => n.id === pendingCheckpoint.nodeId);
-        if (!humanNode) {
-          return sendBadRequest(res, 'Human node not found in workflow');
-        }
-
-        // Use HumanNodeExecutor to process the response
-        const humanExecutor = new HumanNodeExecutor();
-        const resumeResult = await humanExecutor.resume(
-          humanNode,
-          state,
-          { checkpointId, response, data },
-          { executionId, user: req.user }
-        );
-
-        if (resumeResult.status === 'failed') {
-          return sendBadRequest(res, resumeResult.error || 'Failed to process response');
-        }
-
-        // Update execution registry
-        const registry = getExecutionRegistry();
-        registry.clearPendingCheckpoint(executionId);
-
-        // Get the scheduler to determine next nodes based on branch
-        const scheduler = workflowEngine.scheduler;
-        const branch = resumeResult.branch;
-
-        // Build a result object that the scheduler can use for routing
-        const humanResult = {
-          branch,
-          response,
-          ...resumeResult.output
-        };
-
-        // Get next nodes based on the human response branch
-        const nextNodes = scheduler.getNextNodes(humanNode.id, humanResult, workflow, state);
-
-        logger.info('Human checkpoint routing', {
-          component: 'WorkflowRoutes',
-          executionId,
-          humanNodeId: humanNode.id,
-          response,
-          branch,
-          nextNodes
-        });
-
-        // Resume the workflow with the human response and routing info
-        const resumeData = {
-          ...resumeResult.stateUpdates,
-          // Store the human response for edge condition evaluation
-          [`_humanResult_${humanNode.id}`]: humanResult
-        };
-
-        // Update state to mark human node as completed and set next nodes
-        await workflowEngine.stateManager.update(executionId, {
-          completedNodes: [...(state.completedNodes || []), humanNode.id],
-          currentNodes: nextNodes,
-          data: {
-            ...state.data,
-            ...resumeData,
-            nodeResults: {
-              ...(state.data?.nodeResults || {}),
-              [humanNode.id]: humanResult
-            }
-          }
-        });
-
-        const newState = await workflowEngine.resume(
-          executionId,
-          {},
-          {
-            user: req.user,
-            workflow
-          }
-        );
-
-        logger.info('Human checkpoint responded', {
-          component: 'WorkflowRoutes',
-          executionId,
-          checkpointId,
-          response,
-          userId: req.user?.id
-        });
-
-        res.json({
-          success: true,
-          newStatus: newState.status,
-          executionId: newState.executionId
-        });
-      } catch (error) {
-        if (error.code === 'EXECUTION_NOT_FOUND') {
-          return sendNotFound(res, 'Execution');
-        }
-        if (error.code === 'INVALID_STATE_FOR_RESUME') {
-          return sendBadRequest(res, error.message);
-        }
-
-        sendFailedOperationError(res, 'respond to checkpoint', error);
-      }
-    }
-  );
-
-  // ============================================================================
-  // Admin Endpoints
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/admin/workflows:
-   *   get:
-   *     summary: List all workflows (admin view)
-   *     description: |
-   *       Retrieves all workflow definitions including disabled ones.
-   *       Admin access required.
-   *     tags:
-   *       - Workflows
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     responses:
-   *       200:
-   *         description: All workflows
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Admin access required
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/admin/workflows'),
-    checkWorkflowsFeature,
-    adminAuth,
-    async (req, res) => {
-      try {
-        const workflows = await loadWorkflows(true); // Include disabled
-        res.json(workflows);
-      } catch (error) {
-        sendFailedOperationError(res, 'fetch workflows', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/admin/workflows/{id}/toggle:
-   *   post:
-   *     summary: Toggle workflow enabled/disabled status
-   *     description: |
-   *       Toggles the enabled status of a workflow.
-   *       Admin access required.
-   *     tags:
-   *       - Workflows
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Workflow status toggled
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Admin access required
-   *       404:
-   *         description: Workflow not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/admin/workflows/:id/toggle'),
-    checkWorkflowsFeature,
-    adminAuth,
-    async (req, res) => {
-      try {
-        const { id } = req.params;
-
-        // Validate ID for security
-        if (!validateIdForPath(id, 'workflow', res)) {
-          return;
-        }
-
-        // Find workflow file
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-        const filename = await findWorkflowFile(id, workflowsDir);
-
-        if (!filename) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        // Read current workflow
-        const workflowPath = join(workflowsDir, filename);
-        const content = await fs.readFile(workflowPath, 'utf8');
-        const workflow = JSON.parse(content);
-
-        // Toggle enabled status
-        const newEnabledState = workflow.enabled === false;
-        workflow.enabled = newEnabledState;
-
-        // Save updated workflow
-        await atomicWriteJSON(workflowPath, workflow);
-
-        logger.info('Workflow toggled', {
-          component: 'WorkflowRoutes',
-          workflowId: id,
-          enabled: newEnabledState
-        });
-
-        res.json({
-          message: `Workflow ${newEnabledState ? 'enabled' : 'disabled'} successfully`,
-          workflow,
-          enabled: newEnabledState
-        });
-
-        // Refresh workflow cache after successful toggle
-        configCache.refreshWorkflowsCache?.();
-      } catch (error) {
-        sendFailedOperationError(res, 'toggle workflow', error);
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/admin/workflows/executions:
-   *   get:
-   *     summary: List workflow executions (admin view)
-   *     description: |
-   *       Retrieves workflow executions with optional filtering.
-   *       When status=all or a specific status is provided, returns executions
-   *       from the ExecutionRegistry (which includes completed/failed/cancelled).
-   *       Without filters, returns only active (running/paused) executions.
-   *       Admin access required.
-   *     tags:
-   *       - Workflows
-   *       - Admin
-   *       - Execution
-   *     security:
-   *       - adminAuth: []
-   *     parameters:
-   *       - in: query
-   *         name: status
-   *         schema:
-   *           type: string
-   *           enum: [all, running, paused, completed, failed, cancelled]
-   *         description: Filter by execution status. Use 'all' to return all executions.
-   *       - in: query
-   *         name: search
-   *         schema:
-   *           type: string
-   *         description: Search by user ID or workflow name
-   *       - in: query
-   *         name: limit
-   *         schema:
-   *           type: integer
-   *           default: 100
-   *         description: Maximum number of results
-   *       - in: query
-   *         name: offset
-   *         schema:
-   *           type: integer
-   *           default: 0
-   *         description: Number of results to skip
-   *     responses:
-   *       200:
-   *         description: Execution summaries with stats
-   *       401:
-   *         description: Authentication required
-   *       403:
-   *         description: Admin access required
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/admin/workflows/executions'),
-    checkWorkflowsFeature,
-    adminAuth,
-    async (req, res) => {
-      try {
-        const { status, search, limit = 100, offset = 0 } = req.query;
-        const registry = getExecutionRegistry();
-
-        let executions;
-
-        if (status === 'all' || status) {
-          // Get all executions from the registry
-          let allExecutions = Array.from(registry.executions.values()).map(e => ({ ...e }));
-
-          // Apply status filter (unless 'all')
-          if (status && status !== 'all') {
-            allExecutions = allExecutions.filter(e => e.status === status);
-          }
-
-          // Apply search filter on userId or workflowName
-          if (search) {
-            const searchLower = search.toLowerCase();
-            allExecutions = allExecutions.filter(e => {
-              const userId = (e.userId || '').toLowerCase();
-              const workflowName =
-                typeof e.workflowName === 'object'
-                  ? Object.values(e.workflowName).join(' ').toLowerCase()
-                  : (e.workflowName || '').toLowerCase();
-              const workflowId = (e.workflowId || '').toLowerCase();
-              return (
-                userId.includes(searchLower) ||
-                workflowName.includes(searchLower) ||
-                workflowId.includes(searchLower)
-              );
-            });
-          }
-
-          // Sort by startedAt descending (most recent first)
-          allExecutions.sort((a, b) => new Date(b.startedAt) - new Date(a.startedAt));
-
-          // Apply pagination
-          const parsedOffset = parseInt(offset, 10) || 0;
-          const parsedLimit = parseInt(limit, 10) || 100;
-          const total = allExecutions.length;
-          executions = allExecutions.slice(parsedOffset, parsedOffset + parsedLimit);
-
-          // Get stats from registry
-          const stats = registry.getStats();
-
-          res.json({
-            executions,
-            total,
-            stats,
-            offset: parsedOffset,
-            limit: parsedLimit
-          });
-        } else {
-          // Default behavior: return only active executions from WorkflowEngine
-          const activeExecutions = await workflowEngine.listActiveExecutions();
-          const stats = registry.getStats();
-
-          res.json({
-            executions: activeExecutions,
-            total: activeExecutions.length,
-            stats,
-            offset: 0,
-            limit: activeExecutions.length
-          });
-        }
-      } catch (error) {
-        sendFailedOperationError(res, 'fetch executions', error);
-      }
-    }
-  );
-
-  // ============================================================================
-  // Version Control Endpoints
-  // ============================================================================
-
-  /**
-   * @swagger
-   * /api/workflows/{id}/versions:
-   *   get:
-   *     summary: List version history for a workflow
-   *     description: |
-   *       Returns all published version snapshots for a workflow,
-   *       sorted by publish date descending (most recent first).
-   *     tags:
-   *       - Workflows
-   *       - Versions
-   *     security:
-   *       - bearerAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Version list returned successfully
-   *       400:
-   *         description: Invalid workflow ID
-   *       500:
-   *         description: Internal server error
-   */
-  app.get(
-    buildServerPath('/api/workflows/:id/versions'),
-    authRequired,
-    checkWorkflowsFeature,
-    async (req, res) => {
-      const { id } = req.params;
-      if (!validateIdForPath(id, 'workflow', res)) {
-        return;
-      }
-
-      try {
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-        const historyRoot = join(workflowsDir, '.history');
-        // Defense-in-depth: resolve the per-workflow history dir and
-        // assert it stays inside .history. validateIdForPath already
-        // rejects path-traversal characters; this catches symlink/
-        // alternate-encoding bypasses that CodeQL flags.
-        const historyDir = await resolveAndValidatePath(id, historyRoot);
-        if (!historyDir) {
-          return res.status(400).json({ error: 'Invalid workflow ID' });
-        }
-
-        let versions = [];
-        try {
-          // lgtm[js/path-injection] -- `historyDir` is the result of
-          // resolveAndValidatePath(id, historyRoot) above (path.resolve +
-          // startsWith boundary check); `id` was already vetted by
-          // validateIdForPath() which rejects `..`, `/`, `\` and anything
-          // outside [A-Za-z0-9._-].
-          const files = await fs.readdir(historyDir);
-          for (const file of files) {
-            if (!file.endsWith('.json')) continue;
-            try {
-              const filePath = await resolveAndValidatePath(file, historyDir);
-              if (!filePath) continue;
-              // lgtm[js/path-injection] -- `filePath` is the result of
-              // resolveAndValidatePath(file, historyDir) which enforces a
-              // path.resolve + startsWith boundary against historyDir.
-              const content = await fs.readFile(filePath, 'utf8');
-              const data = JSON.parse(content);
-              versions.push({
-                version: data.version,
-                publishedAt: data._publishedAt,
-                publishedBy: data._publishedBy,
-                fileName: file
-              });
-            } catch {
-              // Skip malformed version files
-            }
-          }
-        } catch (err) {
-          if (err.code !== 'ENOENT') throw err;
-          // No history directory yet - return empty
-        }
-
-        // Sort by publish date descending
-        versions.sort((a, b) => new Date(b.publishedAt || 0) - new Date(a.publishedAt || 0));
-
-        res.json({ versions });
-      } catch (error) {
-        logger.error({
-          component: 'workflowRoutes',
-          message: `Failed to list versions: ${error.message}`
-        });
-        res.status(500).json({ error: 'Failed to list versions' });
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/{id}/publish:
-   *   post:
-   *     summary: Publish a workflow version (admin only)
-   *     description: |
-   *       Creates a published snapshot of the current workflow state.
-   *       The snapshot is saved to the version history directory and the
-   *       workflow status is set to 'published'.
-   *     tags:
-   *       - Workflows
-   *       - Versions
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Workflow published successfully
-   *       400:
-   *         description: Invalid workflow ID
-   *       404:
-   *         description: Workflow not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.post(
-    buildServerPath('/api/workflows/:id/publish'),
-    adminAuth,
-    checkWorkflowsFeature,
-    async (req, res) => {
-      const { id } = req.params;
-      if (!validateIdForPath(id, 'workflow', res)) {
-        return;
-      }
-
-      try {
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-
-        // Load current workflow
-        const filename = await findWorkflowFile(id, workflowsDir);
-        if (!filename) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        const workflowPath = join(workflowsDir, filename);
-        const content = await fs.readFile(workflowPath, 'utf8');
-        const workflow = JSON.parse(content);
-
-        // Create snapshot
-        const timestamp = new Date().toISOString().replace(/[:.]/g, '-');
-        const version = workflow.version || '1.0.0';
-        const snapshot = {
-          ...workflow,
-          status: 'published',
-          _publishedAt: new Date().toISOString(),
-          _publishedBy: req.user?.name || req.user?.id || 'unknown'
-        };
-
-        // Save to history (defense-in-depth path resolution)
-        const historyRoot = join(workflowsDir, '.history');
-        const historyDir = await resolveAndValidatePath(id, historyRoot);
-        if (!historyDir) {
-          return res.status(400).json({ error: 'Invalid workflow ID' });
-        }
-        // lgtm[js/path-injection] -- `historyDir` is resolveAndValidatePath's
-        // bounded result for an already validateIdForPath-checked id.
-        await fs.mkdir(historyDir, { recursive: true });
-
-        // version comes from workflow.version which is schema-validated as
-        // semver, so it's a safe filename component. timestamp is also safe.
-        const snapshotFile = join(historyDir, `${version}-${timestamp}.json`);
-        await atomicWriteJSON(snapshotFile, snapshot);
-
-        // Update workflow status
-        workflow.status = 'published';
-        await atomicWriteJSON(workflowPath, workflow);
-
-        logger.info({
-          component: 'workflowRoutes',
-          message: `Published workflow '${id}' version ${version}`,
-          workflowId: id,
-          version
-        });
-
-        res.json({ success: true, version, publishedAt: snapshot._publishedAt });
-      } catch (error) {
-        logger.error({
-          component: 'workflowRoutes',
-          message: `Failed to publish: ${error.message}`
-        });
-        res.status(500).json({ error: 'Failed to publish workflow' });
-      }
-    }
-  );
-
-  /**
-   * @swagger
-   * /api/workflows/{id}/activate/{version}:
-   *   put:
-   *     summary: Activate a specific workflow version (admin only)
-   *     description: |
-   *       Restores a previously published version snapshot as the current
-   *       workflow definition. The snapshot metadata (_publishedAt, _publishedBy)
-   *       is stripped before writing.
-   *     tags:
-   *       - Workflows
-   *       - Versions
-   *       - Admin
-   *     security:
-   *       - adminAuth: []
-   *     parameters:
-   *       - in: path
-   *         name: id
-   *         required: true
-   *         schema:
-   *           type: string
-   *       - in: path
-   *         name: version
-   *         required: true
-   *         schema:
-   *           type: string
-   *     responses:
-   *       200:
-   *         description: Version activated successfully
-   *       400:
-   *         description: Invalid workflow ID or version format
-   *       404:
-   *         description: Workflow or version not found
-   *       500:
-   *         description: Internal server error
-   */
-  app.put(
-    buildServerPath('/api/workflows/:id/activate/:version'),
-    adminAuth,
-    checkWorkflowsFeature,
-    async (req, res) => {
-      const { id } = req.params;
-      if (!validateIdForPath(id, 'workflow', res)) {
-        return;
-      }
-
-      const versionParam = req.params.version;
-      // Strict version validation: safe ID chars only, no `..`, and no leading/trailing punctuation.
-      if (!isValidWorkflowVersion(versionParam)) {
-        return res.status(400).json({ error: 'Invalid version format' });
-      }
-
-      try {
-        const rootDir = getRootDir();
-        const workflowsDir = join(rootDir, 'contents', 'workflows');
-        const historyRoot = join(workflowsDir, '.history');
-        const historyDir = await resolveAndValidatePath(id, historyRoot);
-        if (!historyDir) {
-          return res.status(400).json({ error: 'Invalid workflow ID' });
-        }
-
-        // Find snapshot file by version prefix
-        let snapshotFileName = null;
-        try {
-          // lgtm[js/path-injection] -- `historyDir` is resolveAndValidatePath's
-          // bounded result for a validateIdForPath-checked id.
-          const files = await fs.readdir(historyDir);
-          snapshotFileName = files.find(
-            f => f.startsWith(`${versionParam}-`) && f.endsWith('.json')
-          );
-        } catch {
-          return res.status(404).json({ error: 'No version history found' });
-        }
-
-        if (!snapshotFileName) {
-          return res.status(404).json({ error: `Version ${versionParam} not found` });
-        }
-
-        // Read snapshot (resolve+validate with realpath to prevent symlink traversal)
-        const snapshotPath = await resolveAndValidateRealPath(snapshotFileName, historyDir);
-        if (!snapshotPath) {
-          return res.status(400).json({ error: 'Invalid snapshot path' });
-        }
-        // snapshotPath is a realpath-bounded result constrained to historyDir.
-        const content = await fs.readFile(snapshotPath, 'utf8');
-        const snapshot = JSON.parse(content);
-
-        // Strip metadata
-        delete snapshot._publishedAt;
-        delete snapshot._publishedBy;
-
-        // Write as current workflow
-        const filename = await findWorkflowFile(id, workflowsDir);
-        if (!filename) {
-          return sendNotFound(res, 'Workflow');
-        }
-
-        const workflowPath = join(workflowsDir, filename);
-        await atomicWriteJSON(workflowPath, snapshot);
-
-        logger.info({
-          component: 'workflowRoutes',
-          message: `Activated version ${versionParam} for workflow '${id}'`,
-          workflowId: id,
-          version: versionParam
-        });
-
-        res.json({ success: true, version: versionParam });
-      } catch (error) {
-        logger.error({
-          component: 'workflowRoutes',
-          message: `Failed to activate version: ${error.message}`
-        });
-        res.status(500).json({ error: 'Failed to activate version' });
-      }
-    }
-  );
+  const routeDeps = { ...deps, workflowEngine };
+
+  registerDefinitionRoutes(app);
+  registerExecutionRoutes(app, routeDeps);
+  registerStreamRoutes(app, routeDeps);
+  registerAdminRoutes(app, routeDeps);
+  registerVersionRoutes(app);
 
   return {
     workflowEngine,
@@ -2476,5 +46,11 @@ export default function registerWorkflowRoutes(app, deps = {}) {
   };
 }
 
-// Export helper functions for testing
-export { filterByPermissions, isAdmin, loadWorkflows, findWorkflowFile, validateWorkflow };
+// Re-export helper functions for testing / backward compatibility
+export {
+  filterByPermissions,
+  isAdmin,
+  loadWorkflows,
+  findWorkflowFile,
+  validateWorkflow
+} from './workflowRouteHelpers.js';
